moved surveys into (base-org) directory for layout inheritance
